# Improve Clients query performance