--- a/apps/emqx_gateway/src/emqx_gateway_api_clients.erl
+++ b/apps/emqx_gateway/src/emqx_gateway_api_clients.erl
@@ -56,6 +56,7 @@
 %% internal exports (for client query)
 -export([
     qs2ms/2,
+    run_fuzzy_filter/2,
     format_channel_info/1,
     format_channel_info/2
 ]).
@@ -327,12 +328,7 @@ ms(lifetime, X) ->
 fuzzy_filter_fun([]) ->
     undefined;
 fuzzy_filter_fun(Fuzzy) ->
-    fun(MsRaws) when is_list(MsRaws) ->
-        lists:filter(
-            fun(E) -> run_fuzzy_filter(E, Fuzzy) end,
-            MsRaws
-        )
-    end.
+    {fun ?MODULE:run_fuzzy_filter/2, [Fuzzy]}.
 
 run_fuzzy_filter(_, []) ->
     true;