<!-- Please describe the current behavior and link to a relevant issue. -->
Fixes https://github.com/emqx/emqx/issues/9341

**If your build fails** due to your commit message not passing the build checks, please review the guidelines here: https://github.com/emqx/emqx/blob/master/CONTRIBUTING.md.

## PR Checklist
Please convert it to a draft if any of the following conditions are not met. Reviewers may skip over until all the items are checked:

- [x] Added tests for the changes
- [x] Changed lines covered in coverage report
- [x] Change log has been added to `changes/` dir
- [x] ~For EMQX 4.x: `appup` files updated (execute `scripts/update-appup.sh emqx`)~
- [x] For internal contributor: there is a jira ticket to track this change [EMQX-8137](https://emqx.atlassian.net/browse/EMQX-8137)
- [x] ~If there should be document changes, a PR to emqx-docs.git is sent, or a jira ticket is created to follow up~
- [x] In case of non-backward compatible changes, reviewer should check this item as a write-off, and add details in **Backward Compatibility** section

## Backward Compatibility

1. When performing a fuzzy search, `count` is always returned as `#{meta => #{count := 0}}`, because if we want to get the total number of results, We have to query all rows out and perform a fuzzy matching function each of rows. It's maybe more expensive.

2. Change the `emqx_channel_info` ets table type from `set` to `ordered_set`

## More information

This PR contains 2 fixes:
1. Fix incorrect Swagger Schema for `GET ../clients` and `GET ../subscriptions`.
3. Refactored the implementation of the `emqx_mgmt_api:node_query` and `cluster_query` functions to improve query performance with large data sets.

In the previous implementation, there were two issues that resulted in significant performance overhead:
- **Problem1:** Formatting is performed for each page scanning, whether these pages are returned.
- **Problem2:** Each query traverses the entire dataset, whether it is the first page or last page needed.

## Benchmark
Generate 1M clients into the `emqx_channel_info` table by [emqx-clients-gen script](https://gist.github.com/HJianBo/a7f555e99025702eda0f41da88afad33) and compare the following scenarios

#### Case 1: No conditions
Scripts
```erlang
timer:tc(fun() ->
  emqx_mgmt_api_clients:clients(
    get,
     #{query_string => #{<<"limit">> => 100,<<"page">> => 1}})
end).
```
Report:
| Page   | Before    | Now    |
| ------ | --------- | ------ |
| 1          | 32.304s  | 5.264ms |
| 5000   | 32.935s  | 1.521s |
| 10000 | 31.533s   | 3.074s |

#### Case 2: Performing an exact match
Script:
```erlang
timer:tc(fun() ->
  emqx_mgmt_api_clients:clients(
    get,
    #{query_string => #{<<"limit">> => 100,<<"page">> => 1, <<"username">> => <<"user-999">>}})
 end).
```
Report:
| Page | Before | Now    |
| ---- | ------ | ------ |
| 1    | 1.721s | *3.469s |

`*` It is 2 times more expensive, as the implementation now queries the total and target data separately.

#### Case 3:  Performing a fuzzy search
Script:
```erlang
timer:tc(fun() ->
  emqx_mgmt_api_clients:clients(
    get,
    #{query_string => #{<<"limit">> => 100,<<"page">> => 12,<<"like_username">> => <<"-999">>}})
  end).
```
Report:
| Page | Before | Now    |
| ---- | ------ | ------ |
| 1    | 3.845s | 0.306s |
| 6    | 3.973s | 1.992s |
| 12   | 3.541s | 3.380s |